+++ src/Checkpoint/h5/Wavefield.h
@@ -101,7 +101,7 @@ public:
 	~Wavefield()
 	{ }
 
-	bool init(unsigned int numDofs, unsigned int groupSize = 1);
+	bool init(unsigned long numDofs, unsigned int groupSize = 1);
 
 	void load(double &time, int &timestepWavefield, real* dofs);
