+++ src/Checkpoint/mpio/Wavefield.h
@@ -81,7 +81,7 @@ public:
 	{
 	}
 
-	bool init(unsigned int numDofs, unsigned int groupSize = 1);
+	bool init(unsigned long numDofs, unsigned int groupSize = 1);
 
 	void load(double& time, int& timestepWaveField, real* dofs);
