+++ src/Checkpoint/sionlib/Wavefield.h
@@ -61,7 +61,7 @@ public:
 		: CheckPoint(0x7A389)
 	{}
 
-	bool init(unsigned int numDofs, unsigned int groupSize = 1);
+	bool init(unsigned long numDofs, unsigned int groupSize = 1);
 
 	void load(double &time, int &timestepWavefield, real* dofs);
